refactor: 将碰撞检测算法移动到 collision-detect 中
